layer browser bug fixes
